Moved interceptors, broadcast receiver handlers and json adapters to separate directories. Enabled authentication local datasource.
