+++ app/src/main/java/com/koleff/kare_android/common/network/ApiAuthorizationCallWrapper.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import android.util.Log
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import com.koleff.kare_android.common.Constants
-import com.koleff.kare_android.common.broadcast.RegenerateTokenNotifier
+import com.koleff.kare_android.common.broadcast.handler.RegenerateTokenNotifier
 import com.koleff.kare_android.data.model.response.base_response.KareError
 import com.koleff.kare_android.domain.wrapper.ResultWrapper
 import com.koleff.kare_android.domain.wrapper.ServerResponseData
